day6 jwt

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -9,8 +9,10 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.servlet.http.HttpServletRequest;
 
 import entity.StatusCode;
+import io.jsonwebtoken.Claims;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,12 +21,14 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import util.IdWorker;
 
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -45,6 +49,15 @@ public class UserService {
 
     @Autowired
     private RabbitTemplate rabbitTemplate;
+
+    @Autowired
+    private BCryptPasswordEncoder encoder;
+
+    @Autowired
+    private HttpServletRequest httpServletRequest;
+
+    @Autowired
+    private JwtUtil jwtUtil;
 
     /**
      * 查询全部列表
@@ -99,6 +112,7 @@ public class UserService {
      */
     public void add(User user) {
         user.setId(idWorker.nextId() + "");
+        user.setPassword(encoder.encode(user.getPassword()));
         userDao.save(user);
     }
 
@@ -111,6 +125,7 @@ public class UserService {
             throw new RuntimeException("验证码输入不正确");
         }
         user.setId(idWorker.nextId() + "");
+        user.setPassword(encoder.encode(user.getPassword()));
         user.setFollowcount(0);
         user.setFanscount(0);
         user.setOnline(0L);
@@ -135,6 +150,10 @@ public class UserService {
      * @param id
      */
     public void deleteById(String id) {
+        String token = (String) httpServletRequest.getAttribute("claims_admin");
+        if (token==null||"".equals(token)) {
+            throw new RuntimeException("权限不足!");
+        }
         userDao.deleteById(id);
     }
 
@@ -203,8 +222,17 @@ public class UserService {
         Map<String, String> map = new HashMap<>();
         map.put("mobile", mobile);
         map.put("checkCode", checkCode);
-        rabbitTemplate.convertAndSend("sms", map);
+        // 发送请求到rabbit等消费者消费
+        //rabbitTemplate.convertAndSend("sms", map);
 
         System.out.println("验证码：" + checkCode);
+    }
+
+    public User login(User user){
+        User uerLogin = userDao.findByMobile(user.getMobile());
+        if (uerLogin!=null&&encoder.matches(user.getPassword(),uerLogin.getPassword())) {
+            return uerLogin;
+        }
+        return null;
     }
 }
